+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Enhance: メモリ使用量を削減しました
 - Enhance: ActivityPubアクティビティを送信する際のパフォーマンス向上
 - Enhance: 依存関係の更新
-
+- Enhance: バックエンドプロセス起動前にconfigのymlをjson化するように #16905
 
 ## 2025.11.1