+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Enhance: ActivityPubアクティビティを送信する際のパフォーマンス向上
 - Enhance: 依存関係の更新
 
+
 ## 2025.11.1
 
 ### Client